#316 [refactor] Service 패키지 변경
